backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.23.2
gunicorn==21.2.0
python-dotenv==1.0.0
httpx==0.26.0
python-multipart==0.0.6
pydantic==2.5.3
pydantic-settings==2.1.0
spotipy==2.23.0
anthropic==0.8.1
requests==2.31.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
jinja2==3.1.2
itsdangerous==2.1.2
websockets==12.0
aiofiles==23.2.1